fix: remove customizations for incom tax fields
