+++ india_compliance/gst_india/setup/__init__.py
@@ -14,6 +14,9 @@ from india_compliance.gst_india.constants.custom_fields import (
 )
 from india_compliance.gst_india.setup.property_setters import get_property_setters
 from india_compliance.gst_india.utils import get_data_file_path, toggle_custom_fields
+from india_compliance.patches.post_install.update_e_invoice_fields_and_logs import (
+    delete_custom_fields as _delete_custom_fields,
+)
 
 
 def after_install():
@@ -23,6 +26,11 @@ def after_install():
     set_default_gst_settings()
     set_default_accounts_settings()
     create_hsn_codes()
+
+
+def after_uninstall():
+    delete_custom_fields()
+    delete_property_setters()
 
 
 def create_custom_fields():
@@ -43,6 +51,25 @@ def create_custom_fields():
 def create_property_setters():
     for property_setter in get_property_setters():
         frappe.make_property_setter(property_setter)
+
+
+def delete_custom_fields():
+    _delete_custom_fields(
+        _get_custom_fields_map(
+            CUSTOM_FIELDS,
+            SALES_REVERSE_CHARGE_FIELDS,
+            E_INVOICE_FIELDS,
+            E_WAYBILL_FIELDS,
+        )
+    )
+
+
+def delete_property_setters():
+    for property_setter in get_property_setters():
+        keys_to_update = ["doc_type", "field_name", "property", "value"]
+        # Update the keys to match with the property setter fields
+        filters = dict(zip(keys_to_update, list(property_setter.values())))
+        frappe.db.delete("Property Setter", filters)
 
 
 def create_address_template():
